test(models): split suites and update ci
