+++ .github/workflows/ci.yml
@@ -56,3 +56,25 @@ jobs:
             -destination 'generic/platform=iOS Simulator' \
             -sdk iphonesimulator \
             clean build
+
+  test:
+    name: test
+    runs-on: macos-latest
+    needs: build
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Show Xcode
+        run: |
+          xcodebuild -version
+          xcode-select -p
+
+      - name: Run unit tests (iOS Simulator)
+        run: |
+          xcrun simctl list devices available
+          xcodebuild \
+            -project app/dauphin.xcodeproj \
+            -scheme "dauphin" \
+            -testPlan Models \
+            -destination 'platform=iOS Simulator,name=iPhone 16' \
+            test
