--- App/Tests/Features.xctestplan
@@ -1,24 +1,0 @@
-{
-  "configurations" : [
-    {
-      "id" : "8A660AED-11F0-41EC-9819-51B0619BA7E2",
-      "name" : "Configuration 1",
-      "options" : {
-
-      }
-    }
-  ],
-  "defaultOptions" : {
-    "testTimeoutsEnabled" : true
-  },
-  "testTargets" : [
-    {
-      "target" : {
-        "containerPath" : "container:dauphin.xcodeproj",
-        "identifier" : "188CEB902E838D7000CDD450",
-        "name" : "FeaturesTests"
-      }
-    }
-  ],
-  "version" : 1
-}
